+++ src/main/java/dev/football/sports/api/errors/InvalidScoreException.java
@@ -1,4 +1,4 @@
-package dev.football.sports.domain.exception;
+package dev.football.sports.api.errors;
 
 public class InvalidScoreException extends RuntimeException {
     public InvalidScoreException(String s) {