+++ src/main/java/dev/football/sports/domain/entity/Match.java
@@ -11,11 +11,11 @@ public class Match {
     private final Instant startTime;
 
     public Match(MatchId matchId, Score score, Team home, Team away, Instant startTime) {
-        this.matchId = matchId;
-        this.score = score;
-        this.home = home;
-        this.away = away;
-        this.startTime = startTime;
+        this.matchId = Objects.requireNonNull(matchId);
+        this.score = Objects.requireNonNull(score);
+        this.home = Objects.requireNonNull(home);
+        this.away = Objects.requireNonNull(away);
+        this.startTime = Objects.requireNonNull(startTime);
     }
 
     public MatchId getMatchId() {
@@ -48,5 +48,55 @@ public class Match {
     @Override
     public int hashCode() {
         return Objects.hash(matchId, score, home, away, startTime);
+    }
+
+    public static MatchBuilder builder() {
+        return new MatchBuilder();
+    }
+
+    public static class MatchBuilder {
+        private MatchId matchId;
+        private Score score;
+        private Team home;
+        private Team away;
+        private Instant startTime;
+
+        public MatchBuilder fromEntity(Match match) {
+            matchId = match.matchId;
+            score = match.score;
+            home = match.home;
+            away = match.away;
+            startTime = match.startTime;
+            return this;
+        }
+
+        public MatchBuilder setMatchId(MatchId matchId) {
+            this.matchId = matchId;
+            return this;
+        }
+
+        public MatchBuilder setScore(Score score) {
+            this.score = score;
+            return this;
+        }
+
+        public MatchBuilder setHome(Team home) {
+            this.home = home;
+            return this;
+        }
+
+        public MatchBuilder setAway(Team away) {
+            this.away = away;
+            return this;
+        }
+
+        public MatchBuilder setStartTime(Instant startTime) {
+            this.startTime = startTime;
+            return this;
+        }
+
+        public Match build() {
+            return new Match(matchId, score, home, away, startTime);
+        }
     }
 }