+++ src/main/java/dev/football/sports/domain/service/BoardService.java
@@ -1,5 +1,0 @@
-package dev.football.sports.domain.service;
-
-public interface BoardService {
-
-}