+++ CHANGELOG.md
@@ -9,3 +9,4 @@
     - Essential entities
     - Maven support
     - In-memory storage implementation
+    - DefaultMatchBoard facadewith necessary interface